KITKAT以降で個別のファイル毎にパーミッション要求する場合をSAFUtilsより分離
